+++ source/github-helpers/index.ts
@@ -173,11 +173,3 @@ export const addHotkey = (button: HTMLAnchorElement | HTMLButtonElement | undefi
 		button.dataset.hotkey = [...hotkeys].join(',');
 	}
 };
-
-export function getUserAvatar(username: string, size: number): string {
-	const url = pageDetect.isEnterprise()
-		? `/${username}.png`
-		: `https://avatars.githubusercontent.com/${username}`;
-	// Why use a 2x size: https://github.com/refined-github/refined-github/pull/4973#discussion_r735133613
-	return url + `?size=${size * 2}`;
-}
